+++ Reach/src/utopia/reach/container/multi/stack/Segment.scala
@@ -1,4 +1,4 @@
-package utopia.reach.container
+package utopia.reach.container.multi.stack
 
 import utopia.flow.async.VolatileFlag
 import utopia.flow.datastructure.mutable.ResettableLazy